+++ physics/jacobi_coordinates_test.cpp
@@ -1,7 +1,6 @@
 ﻿
 #include "jacobi_coordinates.hpp"
 
-#include <map>
 #include <vector>
 
 #include "geometry/frame.hpp"
@@ -13,6 +12,7 @@
 namespace principia {
 
 using quantities::si::Kilogram;
+using quantities::si::Metre;
 using testing_utilities::AlmostEquals;
 using testing_utilities::VanishesBefore;
 using ::testing::ElementsAre;
@@ -77,58 +77,6 @@ TEST_F(JacobiCoordinatesTest, Jacobi) {
                           AlmostEquals(-1 * Metre, 2),
                           VanishesBefore(1 * Metre, 1),
                           5 * Metre));
-}
-
-TEST_F(JacobiCoordinatesTest, Hierarchical) {
-
-  // e, i, Ω, ω, and mean anomaly are 0.
-  KeplerianElements<Frame> elements;
-
-  // Invariant: |body_indices[bodies[i]] == i| for all |i|.
-  std::map<not_null<MassiveBody const*>, int> body_indices;
-  std::vector<not_null<MassiveBody const*>> bodies;
-
-  auto const new_body = [&body_indices, &bodies](Mass const& mass) {
-    auto body = make_not_null_unique<MassiveBody>(mass);
-    bodies.emplace_back(body.get());
-    body_indices[body.get()] = body_indices.size();
-    return body;
-  };
-
-  // We construct a system as follows, where the |body_indices|
-  // are, from left to right, 0, 2, 3, 1.
-  // |<1 m>|     |<1 m>|
-  // 2     1     1     2
-  //   |<   7/3 m   >|
-  std::vector<int> left_to_right = {0, 2, 3, 1};
-
-  HierarchicalSystem<Frame> system(new_body(2 * Kilogram));
-  elements.semimajor_axis = 7.0 / 3.0 * Metre;
-  system.Add(new_body(2 * Kilogram), /*parent=*/bodies[0], elements);
-  elements.semimajor_axis = 1 * Metre;
-  system.Add(new_body(1 * Kilogram), /*parent=*/bodies[0], elements);
-  elements.mean_anomaly = π * Radian;
-  system.Add(new_body(1 * Kilogram), /*parent=*/bodies[1], elements);
-
-  auto barycentric_system = system.Get();
-  // primary, closest secondary, furthest secondary, tertiary.
-  std::vector<int> expected_order = {0, 2, 1, 3};
-  for (int i = 0; i < barycentric_system.bodies.size(); ++i) {
-    EXPECT_TRUE(bodies[expected_order[i]] == barycentric_system.bodies[i].get())
-        << i;
-  }
-  std::vector<Length> x_positions;
-  std::transform(barycentric_system.degrees_of_freedom.begin(),
-                 barycentric_system.degrees_of_freedom.end(),
-                 std::back_inserter(x_positions),
-                 [](DegreesOfFreedom<Frame> const& dof) {
-                   return (dof.position() - Frame::origin).coordinates().x;
-                 });
-  EXPECT_THAT(x_positions,
-              ElementsAre(AlmostEquals(-1.5 * Metre, 1),
-                          AlmostEquals(-0.5 * Metre, 2),
-                          1.5 * Metre,
-                          0.5 * Metre));
 }
 
 }  // namespace physics
