+++ physics/physics.vcxproj
@@ -167,6 +167,8 @@
     <ClInclude Include="discrete_trajectory_body.hpp" />
     <ClInclude Include="dynamic_frame.hpp" />
     <ClInclude Include="dynamic_frame_body.hpp" />
+    <ClInclude Include="hierarchical_system.hpp" />
+    <ClInclude Include="hierarchical_system_body.hpp" />
     <ClInclude Include="jacobi_coordinates.hpp" />
     <ClInclude Include="jacobi_coordinates_body.hpp" />
     <ClInclude Include="kepler_orbit.hpp" />
@@ -201,6 +203,7 @@
     <ClCompile Include="degrees_of_freedom_test.cpp" />
     <ClCompile Include="discrete_trajectory_test.cpp" />
     <ClCompile Include="dynamic_frame_test.cpp" />
+    <ClCompile Include="hierarchical_system_test.cpp" />
     <ClCompile Include="jacobi_coordinates_test.cpp" />
     <ClCompile Include="kepler_orbit_test.cpp" />
     <ClCompile Include="resonance_test.cpp" />
