+++ physics/physics.vcxproj.filters
@@ -131,6 +131,12 @@
     <ClInclude Include="jacobi_coordinates_body.hpp">
       <Filter>Source Files</Filter>
     </ClInclude>
+    <ClInclude Include="hierarchical_system.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="hierarchical_system_body.hpp">
+      <Filter>Source Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="degrees_of_freedom_test.cpp">
@@ -173,6 +179,9 @@
       <Filter>Test Files</Filter>
     </ClCompile>
     <ClCompile Include="jacobi_coordinates_test.cpp">
+      <Filter>Test Files</Filter>
+    </ClCompile>
+    <ClCompile Include="hierarchical_system_test.cpp">
       <Filter>Test Files</Filter>
     </ClCompile>
   </ItemGroup>
